refactor parse color
