+++ components/lib/parseColor.js
